resolve common dirs
