+++ .flowconfig
@@ -11,3 +11,6 @@ flow-typed
 module.file_ext=.js
 module.file_ext=.jsx
 module.ignore_non_literal_requires=true
+
+module.system.node.resolve_dirname=node_modules
+module.system.node.resolve_dirname=src
