+++ package.json
@@ -6,8 +6,9 @@
     "scripts": {
         "build": "rm -rf dist && NODE_ENV=production webpack",
         "flow": "flow",
-        "test": "echo \"Error: no test specified\" && exit 1",
+        "test": "echo \"Write tests!\"",
         "lint": "eslint . --ext .js --quiet",
+        "validate": "yarn lint && yarn flow && yarn test",
         "fix": "eslint . --ext .js --fix --quiet",
         "dev": "webpack-dev-server --hot --env.browser"
     },
@@ -52,6 +53,7 @@
         "eslint": "^4.9.0",
         "eslint-config-airbnb": "^16.1.0",
         "eslint-config-prettier": "^2.6.0",
+        "eslint-import-resolver-webpack": "^0.8.4",
         "eslint-plugin-flow-header": "^0.2.0",
         "eslint-plugin-flowtype": "^2.44.0",
         "eslint-plugin-import": "^2.7.0",
@@ -94,6 +96,9 @@
             "prettier/flowtype",
             "prettier/react"
         ],
+        "settings": {
+            "import/resolver": "webpack"
+        },
         "plugins": [
             "flow-header",
             "flowtype",
